requirements.txt
annoy==1.17.0
asgiref==3.5.0
attrs==19.1.0
backcall==0.1.0
backports.entry-points-selectable==1.1.0
bleach==3.1.0
certifi==2021.5.30
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.2
Django==3.2.11
djangorestframework==3.13.1
entrypoints==0.3
filelock==3.5.1
importlib-metadata==4.6.1
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.6.0
nbformat==4.4.0
nose==1.3.7
notebook==6.0.0
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
platformdirs==2.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.9.3
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.3
scipy==1.3.1
Send2Trash==1.5.0
six==1.16.0
sqlparse==0.4.2
sympy==1.4
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typing_extensions==4.0.1
virtualenv==20.13.1
virtualenv-clone==0.5.5
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.5.0
